requirements.txt
streamlit==1.31.1
PyPDF2==3.0.1
langchain==0.2.6
langchain-core==0.2.10
langchain-google-genai==1.0.6
faiss-cpu==1.8.0
python-dotenv==1.0.1
google-generativeai==0.5.4
pydantic<3